Move VerificationModal to App/index
