+++ front/app/components/PostShowComponents/Comments/CommentingInitiativeDisabled.tsx
@@ -13,7 +13,7 @@ import messages from './messages';
 import { FormattedMessage } from 'utils/cl-intl';
 
 // events
-import { openVerificationModal } from 'components/Verification/verificationModalEvents';
+import { openVerificationModal } from 'containers/App/VerificationModal/verificationModalEvents';
 import { openSignUpInModal } from 'components/SignUpIn/events';
 
 // styling
