+++ front/app/components/VoteControl/index.tsx
@@ -34,7 +34,7 @@ import {
 
 // utils
 import { openSignUpInModal } from 'components/SignUpIn/events';
-import { openVerificationModal } from 'components/Verification/verificationModalEvents';
+import { openVerificationModal } from 'containers/App/VerificationModal/verificationModalEvents';
 
 // style
 import styled from 'styled-components';
