+++ front/app/containers/App/VerificationModal/VerificationSuccess.tsx
