+++ front/app/containers/App/VerificationModal/messages.ts
@@ -1,0 +1,108 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  // verifyYourIdentity: {
+  //   id: 'app.components.VerificationModal.verifyYourIdentity',
+  //   defaultMessage: 'Verify your identity',
+  // },
+  // andUnlockYourCitizenPotential: {
+  //   id: 'app.components.VerificationModal.andUnlockYourCitizenPotential',
+  //   defaultMessage: 'and unlock your citizen potential',
+  // },
+  // toParticipateInThisProject: {
+  //   id: 'app.components.VerificationModal.toParticipateInThisProject',
+  //   defaultMessage: 'to participate in this project',
+  // },
+  // showHelp: {
+  //   id: 'app.components.VerificationModal.showHelp',
+  //   defaultMessage: 'Where can I find my ID Serial Number?',
+  // },
+  // or: {
+  //   id: 'app.components.VerificationModal.or',
+  //   defaultMessage: 'or',
+  // },
+  // takenFormError: {
+  //   id: 'app.components.VerificationModal.takenFormError',
+  //   defaultMessage: 'Already taken.',
+  // },
+  // emptyFieldError: {
+  //   id: 'app.components.VerificationModal.emptyFieldError',
+  //   defaultMessage: 'This field cannot be empty.',
+  // },
+  // noMatchFormError: {
+  //   id: 'app.components.VerificationModal.noMatchFormError',
+  //   defaultMessage: 'No match was found.',
+  // },
+  // notEntitledFormError: {
+  //   id: 'app.components.VerificationModal.notEntitledFormError',
+  //   defaultMessage: 'Not entitled.',
+  // },
+  // invalidCardIdError: {
+  //   id: 'app.components.VerificationModal.invalidCardIdError',
+  //   defaultMessage: 'This id is not valid.',
+  // },
+  // somethingWentWrongError: {
+  //   id: 'app.components.VerificationModal.somethingWentWrongError',
+  //   defaultMessage: "We can't verify you because something went wrong",
+  // },
+  // submit: {
+  //   id: 'app.components.VerificationModal.submit',
+  //   defaultMessage: 'Submit',
+  // },
+  // cancel: {
+  //   id: 'app.components.VerificationModal.cancel',
+  //   defaultMessage: 'Cancel',
+  // },
+  // back: {
+  //   id: 'app.components.VerificationModal.back',
+  //   defaultMessage: 'Back',
+  // },
+  // userVerifiedTitle: {
+  //   id: 'app.components.VerificationModal.userVerifiedTitle',
+  //   defaultMessage: 'You’re now verified !',
+  // },
+  // userVerifiedSubtitle: {
+  //   id: 'app.components.VerificationModal.userVerifiedSubtitle',
+  //   defaultMessage:
+  //     'Your identity has been verified. You’re now a full member of the community on this platform.',
+  // },
+  errorTitle: {
+    id: 'app.components.VerificationModal.errorTitle',
+    defaultMessage: 'There was an issue with the verification of your account',
+  },
+  errorTakenSubtitle: {
+    id: 'app.components.VerificationModal.errorTakenSubtitle',
+    defaultMessage:
+      'Verification failed because another account is already using this identity.',
+  },
+  errorNotEntitledSubtitle: {
+    id: 'app.components.VerificationModal.errorNotEntitledSubtitle',
+    defaultMessage:
+      "Your identity has been found, but it doesn't have enough civil rights to be considered verified",
+  },
+  errorGenericSubtitle: {
+    id: 'app.components.VerificationModal.errorGenericSubtitle',
+    defaultMessage:
+      'There was an unkown issue with the verification of your account, please try again.',
+  },
+  // participationConditions: {
+  //   id: 'app.components.VerificationModal.participationConditions',
+  //   defaultMessage: 'Participation conditions',
+  // },
+  // peopleMatchingConditions: {
+  //   id: 'app.components.VerificationModal.peopleMatchingConditions',
+  //   defaultMessage: 'People matching the following conditions:',
+  // },
+  // verifyNow: {
+  //   id: 'app.components.VerificationModal.verifyNow',
+  //   defaultMessage: 'Verify now',
+  // },
+  // skipThisStep: {
+  //   id: 'app.components.VerificationModal.skipThisStep',
+  //   defaultMessage: 'Skip this step',
+  // },
+  // close: {
+  //   id: 'app.components.VerificationModal.close',
+  //   defaultMessage: 'Close',
+  // },
+});
