+++ front/app/containers/ProjectsShowPage/shared/poll/index.tsx
@@ -29,7 +29,7 @@ import { MessageDescriptor } from 'react-intl';
 
 // events
 import { openSignUpInModal } from 'components/SignUpIn/events';
-import { openVerificationModal } from 'components/Verification/verificationModalEvents';
+import { openVerificationModal } from 'containers/App/VerificationModal/verificationModalEvents';
 
 import styled from 'styled-components';
