+++ front/app/services/singleSignOn.ts
@@ -2,8 +2,7 @@ import { AUTH_PATH } from 'containers/App/constants';
 import { ISignUpInMetaData, TSignUpInError } from 'components/SignUpIn';
 import { stringify } from 'qs';
 import { omitBy, isNil } from 'lodash-es';
-import { isProjectContext } from 'components/Verification/verificationModalEvents';
-
+import { isProjectContext } from 'containers/App/VerificationModal/verificationModalEvents';
 export interface SSOProviderMap {
   azureactivedirectory: 'azureactivedirectory';
   facebook: 'facebook';
